refactor(package-manager): decompose install modules
