+++ pnpm/crates/patching/Cargo.toml
@@ -11,11 +11,15 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pacquet-git-fetcher      = { workspace = true }
+pacquet-package-manifest = { workspace = true }
+
 derive_more = { workspace = true }
 diffy       = { workspace = true }
 indexmap    = { workspace = true }
 miette      = { workspace = true }
 node-semver = { workspace = true }
+serde_json  = { workspace = true }
 sha2        = { workspace = true }
 
 [dev-dependencies]
